somewhat

+++ java/src/main/java/com/angrysurfer/midi/model/Sample.java
@@ -19,7 +19,7 @@ public class Sample extends Strike {
     }
 
     @Override
-    public void onTick(long tick, int bar) {
+    public void onTick(long tick, long bar) {
 
     }
 }
